move combiner and post-processor types
