+++ src/vivarium/framework/values/__init__.py
@@ -12,18 +12,17 @@ For more information about when and how you should use pipelines in your
 simulations, see the value system :ref:`concept note <values_concept>`.
 
 """
-from vivarium.framework.values.combiners import list_combiner, replace_combiner
+from vivarium.framework.values.combiners import ValueCombiner, list_combiner, replace_combiner
 from vivarium.framework.values.exceptions import DynamicValueError
 from vivarium.framework.values.manager import ValuesInterface, ValuesManager
 from vivarium.framework.values.pipeline import (
     MissingValueSource,
     Pipeline,
-    PostProcessor,
-    ValueCombiner,
     ValueModifier,
     ValueSource,
 )
 from vivarium.framework.values.post_processors import (
+    PostProcessor,
     rescale_post_processor,
     union_post_processor,
 )
